backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-genai>=1.46.0
pydantic>=2.11.7
python-dotenv>=1.1.0
elevenlabs>=1.0.0
imageio-ffmpeg>=0.4.9
httpx>=0.24.0

